% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_leggett_app.R
\name{launch_leggett_app}
\alias{launch_leggett_app}
\title{Launch the Leggett+ Shiny application}
\usage{
launch_leggett_app()
}
\value{
Invisibly returns `NULL`. Called for its side effect of starting a
  Shiny application.
}
\description{
Launch the Leggett+ Shiny application
}
\examples{
if (interactive()) {
  launch_leggett_app()
}
}
