% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leggett_ode.R
\name{leggett_ode}
\alias{leggett_ode}
\title{Leggett+ differential equation system}
\usage{
leggett_ode(parameters)
}
\arguments{
\item{parameters}{Static parameters produced by [leggett_base_parameters()].}
}
\value{
Function of time, state, and stage parameters.
}
\description{
Returns an ODE function suitable for `deSolve::ode`.
}
\keyword{internal}
