% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeggettPlus-package.R
\docType{package}
\name{LeggettPlus-package}
\alias{LeggettPlus}
\alias{LeggettPlus-package}
\title{LeggettPlus}
\description{
LeggettPlus provides an open-source, reproducible implementation of the
Leggett+ physiologically based pharmacokinetic (PBPK) model for lead.
}
\details{
This package was created by the Office of Environmental Health Hazard
Assessment (OEHHA), an office within the State of California's
Environmental Protection Agency (CalEPA). It is intended for research and
exposure-scenario exploration only. Outputs should be interpreted within
the context of model assumptions, parameter uncertainty, and available
toxicokinetic data. This package should not be used and is not intended to
be used for clinical diagnosis or individual medical decision-making.
OEHHA provides no warranty for this package, either express or implied,
including but not limited to the implied warranties of merchantability or
fitness for a particular purpose; it is provided "AS IS." OEHHA disclaims
any liability resulting from the use of this package.

Physiological parameters (body weight, hematocrit, blood-volume scaling,
and RBC binding capacity) are derived from adult male / male-dominated
occupational-worker data (Leggett, 1993, and subsequent updates). There is
no female-specific parameter set implemented (e.g. a lower hematocrit
range or pregnancy-related blood-volume expansion); results should not be
assumed to represent female physiology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OEHHA-NTES/Leggett_Plus}
  \item Report bugs at \url{https://github.com/OEHHA-NTES/Leggett_Plus/issues}
}

}
\author{
\strong{Maintainer}: Scott Coffin \email{scott.l.coffin@gmail.com} (\href{https://orcid.org/0000-0002-7035-1282}{ORCID})

Authors:
\itemize{
  \item Scott Coffin \email{scott.l.coffin@gmail.com} (\href{https://orcid.org/0000-0002-7035-1282}{ORCID})
  \item Hsing-Chieh Lin \email{hclin@tamu.edu}
}

Other contributors:
\itemize{
  \item New Toxicology Evaluations Section, Office of Environmental Health Hazard Assessment \email{NTES@oehha.ca.gov} [copyright holder, funder]
}

}
